Move non-API classes to core
