+++ log4j-core/src/main/java/org/apache/logging/log4j/core/util/ContextDataProvider.java
@@ -22,11 +22,26 @@ import org.apache.logging.log4j.util.StringMap;
 
 /**
  * Source of context data to be added to each log event.
- * @deprecated Use ContextDataProvider from Log4j API from 2.24.0.
  */
-@Deprecated
-public interface ContextDataProvider extends org.apache.logging.log4j.spi.ContextDataProvider {
+public interface ContextDataProvider {
 
+    /**
+     * Returns the key for a value from the context data.
+     * @param key the key to locate.
+     * @return the value or null if it is not found.
+     */
+    default String get(String key) {
+        return null;
+    }
+
+    /**
+     * Returns a Map containing context data to be injected into the event or null if no context data is to be added.
+     * <p>
+     *     Thread-safety note: The returned object can safely be passed off to another thread: future changes in the
+     *     underlying context data will not be reflected in the returned object.
+     * </p>
+     * @return A Map containing the context data or null.
+     */
     Map<String, String> supplyContextData();
 
     /**
@@ -41,5 +56,36 @@ public interface ContextDataProvider extends org.apache.logging.log4j.spi.Contex
     @Deprecated
     default StringMap supplyStringMap() {
         return new JdkMapAdapterStringMap(supplyContextData(), true);
+    }
+
+    /**
+     * Returns the number of items in this context.
+     * @return the number of items in the context.
+     */
+    default int size() {
+        Map<String, String> contextMap = supplyContextData();
+        return contextMap != null ? contextMap.size() : 0;
+    }
+
+    /**
+     * Add all the keys in the current context to the provided Map.
+     * @param map the StringMap to add the keys and values to.
+     */
+    default void addAll(Map<String, String> map) {
+        Map<String, String> contextMap = supplyContextData();
+        if (contextMap != null) {
+            map.putAll(contextMap);
+        }
+    }
+
+    /**
+     * Add all the keys in the current context to the provided StringMap.
+     * @param map the StringMap to add the keys and values to.
+     */
+    default void addAll(StringMap map) {
+        Map<String, String> contextMap = supplyContextData();
+        if (contextMap != null) {
+            contextMap.forEach(map::putValue);
+        }
     }
 }
